Rename cssStatementHasBlock to hasBlock
