+++ src/rules/block-opening-brace-newline-after/index.js
@@ -1,6 +1,6 @@
 import {
   blockString,
-  cssStatementHasBlock,
+  hasBlock,
   cssStatementHasEmptyBlock,
   cssStatementStringBeforeBlock,
   rawNodeString,
@@ -39,7 +39,7 @@ export default function (expectation) {
     function check(statement) {
 
       // Return early if blockless or has an empty block
-      if (!cssStatementHasBlock(statement) || cssStatementHasEmptyBlock(statement)) { return }
+      if (!hasBlock(statement) || cssStatementHasEmptyBlock(statement)) { return }
 
       // Allow an end-of-line comment x spaces after the brace
       const firstNode = statement.first
