+++ src/rules/block-opening-brace-space-after/index.js
@@ -1,6 +1,6 @@
 import {
   blockString,
-  cssStatementHasBlock,
+  hasBlock,
   cssStatementHasEmptyBlock,
   cssStatementStringBeforeBlock,
   report,
@@ -42,7 +42,7 @@ export default function (expectation) {
 
     function check(statement) {
       // Return early if blockless or has an empty block
-      if (!cssStatementHasBlock(statement) || cssStatementHasEmptyBlock(statement)) { return }
+      if (!hasBlock(statement) || cssStatementHasEmptyBlock(statement)) { return }
 
       checker.after({
         source: blockString(statement),
