+++ src/rules/max-nesting-depth/index.js
@@ -1,6 +1,6 @@
 import _ from "lodash"
 import {
-  cssStatementHasBlock,
+  hasBlock,
   optionsHaveIgnored,
   ruleMessages,
   report,
@@ -32,7 +32,7 @@ export default function (max, options) {
     root.walkAtRules(checkStatement)
 
     function checkStatement(statement) {
-      if (!cssStatementHasBlock(statement)) { return }
+      if (!hasBlock(statement)) { return }
       const depth = nestingDepth(statement)
       if (depth > max) {
         report({
