+++ src/rules/number-max-precision/index.js
@@ -2,7 +2,7 @@ import { isNumber } from "lodash"
 import execall from "execall"
 import {
   blurComments,
-  cssStatementHasBlock,
+  hasBlock,
   cssStatementStringBeforeBlock,
   report,
   ruleMessages,
@@ -30,7 +30,7 @@ export default function (precision) {
     })
 
     root.walkAtRules(atRule => {
-      const source = (cssStatementHasBlock(atRule))
+      const source = (hasBlock(atRule))
         ? cssStatementStringBeforeBlock(atRule, { noBefore: true })
         : atRule.toString()
       check(source, atRule)
