+++ src/utils/hasBlock.js
