+++ src/utils/index.js
@@ -4,7 +4,7 @@ export { default as configurationError } from "./configurationError"
 export { default as functionArgumentsSearch } from "./functionArgumentsSearch"
 export { default as cssNodeContextLookup } from "./cssNodeContextLookup"
 export { default as blockString } from "./blockString"
-export { default as cssStatementHasBlock } from "./cssStatementHasBlock"
+export { default as hasBlock } from "./hasBlock"
 export { default as cssStatementHasEmptyBlock } from "./cssStatementHasEmptyBlock"
 export { default as cssStatementStringBeforeBlock } from "./cssStatementStringBeforeBlock"
 export { default as cssWordIsVariable } from "./cssWordIsVariable"
